requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pyyaml